Merge in 2017 code.
I decided not to do a git merge here. I could've, in principle. But I
really do not know what would have happened if there were changes in the
old history of the 2017 repo that touched the framework code paths, and
I don't really want to find out. If we have questions we can go back to
the 2017 repo.