Swapped hardware-based limit switches for software-based ones
Due to complexities with the limit switches installed in the Talon SRXs, we want to use software-based limit switches when possible. This entails stopping movement when a limit switch has been activated and the StrafeAxis is still moving in the direction of the activated limit switch. Hardware-based limit switch logic is still there, if needed.