Began updating Pathfinder for 2019
Added support for JacisNonsense's implementation of Pathfinder for FRC (mainly used in reading the path files). Please note that a lot of this code is experimental and still needs testing. The robot code may crash due to a NullPointerException, for example, because the file paths used in PathNameConstants.java need to be rewritten. Also, the logic for finding left and right files needs to be written in as opposed to hard coding the file names in.